+++ packages/account-postgres-sink-service/src/utils/handleAccountWebhook.ts
@@ -1,6 +1,6 @@
 import * as anchor from "@coral-xyz/anchor";
 import { PublicKey } from "@solana/web3.js";
-import deepEqual from "deep-equal";
+import deepEqual from "fast-deep-equal";
 import { FastifyInstance } from "fastify";
 import _omit from "lodash/omit";
 import { Sequelize, Transaction } from "sequelize";
@@ -90,10 +90,13 @@ export const handleAccountWebhook = async ({
       );
 
       if (isDelete) {
-        await model.destroy({
-          where: { address: account.pubkey },
-          transaction: t,
-        });
+        let ignoreDelete = accounts.find(acc => acc.type == accName)?.ignore_deletes;
+        if (!ignoreDelete) {
+          await model.destroy({
+            where: { address: account.pubkey },
+            transaction: t,
+          });
+        }
       } else {
         sanitized = {
           refreshed_at: new Date().toISOString(),