+++ packages/helium-sub-daos-sdk/src/resolvers.ts
@@ -12,6 +12,7 @@ import { EPOCH_LENGTH, PROGRAM_ID } from "./constants";
 import { init as initNftProxy } from "@helium/nft-proxy-sdk";
 import { init as initHsd } from "./init";
 import { daoEpochInfoKey, subDaoEpochInfoKey } from "./pdas";
+import { min } from "bn.js";
 
 const THREAD_PID = new PublicKey(
   "CLoCKyJ6DXBJqqu2VWx9RLbgnwwR6BMHHuyasVmfMzBh"
@@ -116,6 +117,22 @@ export const subDaoEpochInfoResolver = resolveIndividual(
         return key;
       }
     }
+    if (path[path.length - 1] === "prevSubDaoEpochInfo" && args && args[0] && args[0].epoch) {
+      const unixTime = args[0].epoch.toNumber() * EPOCH_LENGTH;
+      const subDao = get(accounts, [
+        ...path.slice(0, path.length - 1),
+        "subDao",
+      ]) as PublicKey;
+      if (subDao) {
+        const [key] = await subDaoEpochInfoKey(
+          subDao,
+          unixTime - EPOCH_LENGTH,
+          PROGRAM_ID
+        );
+
+        return key;
+      }
+    }
   }
 );
 
@@ -154,7 +171,7 @@ export const closingTimeEpochInfoResolver = resolveIndividual(
       if (positionAcc && (proxyConfigAcc || delegatedPositionAcc)) {
         const expirationTs =
           !delegatedPositionAcc || delegatedPositionAcc.expirationTs.isZero()
-            ? proxyConfigAcc?.seasons.find((s) =>
+            ? proxyConfigAcc?.seasons?.reverse().find((s) =>
                 new BN(now.toString()).gte(s.start)
               )?.end || positionAcc.lockup.endTs
             : delegatedPositionAcc.expirationTs;
@@ -180,6 +197,8 @@ export const genesisEndEpochInfoResolver = resolveIndividual(
   async ({ provider, path, accounts }) => {
     if (path[path.length - 1] === "genesisEndSubDaoEpochInfo") {
       const program = await init(provider as AnchorProvider, VSR_PROGRAM_ID);
+      const hsdProgram = await initHsd(provider as AnchorProvider);
+      const nftProxyProgram = await initNftProxy(provider as AnchorProvider);
 
       const subDao = get(accounts, [
         ...path.slice(0, path.length - 1),
@@ -195,15 +214,35 @@ export const genesisEndEpochInfoResolver = resolveIndividual(
       ]) as PublicKey;
       const positionAcc = position && await program.account.positionV0.fetch(position);
       const registrarAcc = registrar && await program.account.registrar.fetch(registrar);
+      const proxyConfig = get(accounts, [
+        ...path.slice(0, path.length - 1),
+        "proxyConfig",
+      ]) as PublicKey;
+      const delegatedPosition = get(accounts, [
+        ...path.slice(0, path.length - 1),
+        "delegatedPosition",
+      ]) as PublicKey;
+      const delegatedPositionAcc =
+        delegatedPosition &&
+        (await hsdProgram.account.delegatedPositionV0.fetchNullable(
+          delegatedPosition
+        ));
+      const proxyConfigAcc = proxyConfig && await nftProxyProgram.account.proxyConfigV0.fetch(proxyConfig);
       if (positionAcc && registrarAcc) {
-        const currTs = Number(await getSolanaUnixTimestamp(provider)) + registrarAcc.timeOffset.toNumber();
-        const ts =
-          positionAcc.genesisEnd.toNumber() < currTs
-            ? positionAcc.lockup.endTs.toNumber()
-            : positionAcc.genesisEnd;
+        const now =
+          Number(await getSolanaUnixTimestamp(provider)) +
+          registrarAcc.timeOffset.toNumber();
+        const seasonEnd = proxyConfigAcc?.seasons
+          ?.reverse()
+          .find((s) => new BN(now.toString()).gte(s.start))?.end;
+        const expirationTs =
+          !delegatedPositionAcc || delegatedPositionAcc.expirationTs.isZero()
+            ? seasonEnd || positionAcc.lockup.endTs
+            : delegatedPositionAcc.expirationTs;
+        const epochTs = positionAcc.genesisEnd.lte(new BN(now)) ? min(positionAcc.lockup.endTs, expirationTs) : positionAcc.genesisEnd;
         const [key] = await subDaoEpochInfoKey(
           subDao,
-          ts
+          epochTs
         );
 
         return key;
@@ -265,6 +304,12 @@ export const heliumSubDaosResolvers = combineResolvers(
     account: "positionTokenAccount",
     mint: "mint",
     owner: "positionAuthority",
+  }),
+  ataResolver({
+    instruction: "extendExpirationTsV0",
+    account: "positionTokenAccount",
+    mint: "mint",
+    owner: "authority",
   }),
   resolveIndividual(async ({ args, path, accounts }) => {
     if (path[path.length - 1] == "clockwork") {