+++ programs/helium-sub-daos/src/instructions/delegation/claim_rewards_v1.rs
@@ -32,6 +32,7 @@ pub struct ClaimRewardsV1<'info> {
     has_one = registrar,
   )]
   pub position: Box<Account<'info, PositionV0>>,
+  #[account(mut)]
   pub mint: Box<Account<'info, Mint>>,
   #[account(
     token::mint = mint,
@@ -137,8 +138,12 @@ pub fn handler(ctx: Context<ClaimRewardsV1>, args: ClaimRewardsArgsV0) -> Result
     }
   }
 
-  let delegated_vehnt_at_epoch =
-    position.voting_power(voting_mint_config, ctx.accounts.dao_epoch_info.start_ts())?;
+  let epoch_start_ts = ctx.accounts.dao_epoch_info.start_ts();
+  let delegated_vehnt_at_epoch = if delegated_position.expiration_ts > epoch_start_ts {
+    position.voting_power(voting_mint_config, epoch_start_ts)?
+  } else {
+    0
+  };
 
   msg!("Staked {} veHNT at start of epoch with {} total veHNT delegated to dao and {} total rewards to dao",
     delegated_vehnt_at_epoch,
@@ -187,7 +192,7 @@ pub fn handler(ctx: Context<ClaimRewardsV1>, args: ClaimRewardsArgsV0) -> Result
       .dao
       .recent_proposals
       .iter()
-      .filter(|p| p.proposal == Pubkey::default())
+      .filter(|p| p.proposal != Pubkey::default())
       .count()
       < 2;