+++ utils/vehnt/src/cli/delegated.rs
@@ -112,6 +112,9 @@ impl Delegated {
     let positions_with_delegations: Vec<FullPosition> = delegated_positions
       .iter()
       .zip(positions_raw)
+      .filter(|((_, delegated_position), position)| {
+        position.is_some()
+      })
       .map(|((pubkey, delegated_position), position)| {
         let position_unwrapped = position.unwrap();
         let mut data = position_unwrapped.data.as_slice();
@@ -165,10 +168,12 @@ impl Delegated {
     epoch_infos_by_subdao_and_epoch
       .insert(Pubkey::from_str(MOBILE_SUBDAO).unwrap(), HashMap::new());
     for info in infos {
+        if info.1.sub_dao == Pubkey::from_str(IOT_SUBDAO).unwrap() || info.1.sub_dao == Pubkey::from_str(MOBILE_SUBDAO).unwrap() {
       epoch_infos_by_subdao_and_epoch
-        .get_mut(&info.1.sub_dao)
-        .unwrap()
-        .insert(info.1.epoch, info);
+          .get_mut(&info.1.sub_dao)
+          .unwrap()
+          .insert(info.1.epoch, info);
+      }
     }
     let mut new_epoch_infos_by_subdao_and_epoch: HashMap<
       Pubkey,